Add in PNPM stuff, resolve conflicts, rush generate
